Get example program working with new API
